Refactoring code
Function's names and file structure was changed.
