+++ src/App.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import s from './App.module.css'
-import InputContainer from "./components/InputContainer";
+import { InputContainer } from "./components/Input/InputContainer";
 import './App.module.css';
 
 const App = () => {
