+++ src/components/Input/InputContainer.type.ts
@@ -1,0 +1,5 @@
+export type MessageType = {
+    codeStatus: number,
+    message: string,
+    status: string,
+};
